+++ app/controllers/admin/departments_controller.rb
@@ -1,16 +1,29 @@
 class Admin::DepartmentsController < ApplicationController
   before_action :authenticate_user!
-  before_action :require_admin, except: [:index, :show]
 
   def index
+    # Get departments
     @departments = Department.all.paginate(
       :page => params[:page],
       :per_page => 10
     )
+    # Check with pundit if the user has permission
+    authorize @departments
+  end
+
+  def new
+    # Create new instance of department
+    @department = Department.new
+    # Check with pundit if able to create new
+    authorize @department
   end
 
   def create
+    # Create new instance from params
     @department = Department.new(department_params)
+    # Check with pundit if the user has permission
+    authorize @department
+    # Survived so save department
     if @department.save
       flash[:notice] = 'Department successfully created'
       redirect_to admin_departments_path
@@ -19,16 +32,19 @@ class Admin::DepartmentsController < ApplicationController
     end
   end
 
-  def new
-    @department = Department.new
-  end
-
   def edit
+    # Get department
     @department = Department.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @department
   end
 
   def update
+    # Get department
     @department = Department.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @department
+    # Survived so update
     if @department && @department.update(department_params)
       flash[:notice] = 'Department successfully updated'
       redirect_to admin_departments_path
@@ -38,7 +54,11 @@ class Admin::DepartmentsController < ApplicationController
   end
 
   def destroy
+    # Get deparment
     @department = Department.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @department
+    # Survived so destroy
     if @department && @department.destroy
       flash[:notice] = 'Departmnt successfully removed'
     else
