+++ app/controllers/admin/qualifiers_controller.rb
@@ -1,16 +1,29 @@
 class Admin::QualifiersController < ApplicationController
   before_action :authenticate_user!
-  before_action :require_admin, except: [:index, :show]
 
   def index
+    # Get qualifiers
     @qualifiers = Qualifier.all.paginate(
       :page => params[:page],
       :per_page => 10
     )
+    # Check with pundit if the user has permission
+    authorize @qualifiers
+  end
+
+  def new
+    # Create new qualifier instance
+    @qualifier = Qualifier.new
+    # Check with pundit if the user has permission
+    authorize @qualifier
   end
 
   def create
+    # Create qualifer from params
     @qualifier = Qualifier.new(qualifier_params)
+    # Check with pundit if the user has permission
+    authorize @qualifier
+    # Survived so save
     if @qualifier.save
       flash[:notice] = 'Qualifier successfully created'
       redirect_to admin_qualifiers_path
@@ -19,16 +32,19 @@ class Admin::QualifiersController < ApplicationController
     end
   end
 
-  def new
-    @qualifier = Qualifier.new
-  end
-
   def edit
+    # Get qualifier
     @qualifier = Qualifier.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @qualifier
   end
 
   def update
+    # Get qualifier
     @qualifier = Qualifier.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @qualifier
+    # Survived so update
     if @qualifier && @qualifier.update(qualifier_params)
       flash[:notice] = 'Qualifier successfully updated'
       redirect_to admin_qualifiers_path
@@ -39,6 +55,9 @@ class Admin::QualifiersController < ApplicationController
 
   def destroy
     @qualifier = Qualifier.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @qualifier
+    # Survived so destroy
     if @qualifier && @qualifier.destroy
       flash[:notice] = 'Qualifier successfully removed'
     else
