+++ app/controllers/admin/sale_types_controller.rb
@@ -1,16 +1,29 @@
 class Admin::SaleTypesController < ApplicationController
   before_action :authenticate_user!
-  before_action :require_admin, except: [:index, :show]
 
   def index
+    # Get sale types
     @saletypes = SaleType.all.paginate(
       :page => params[:page],
       :per_page => 10
     )
+    # Check with pundit if the user has permission
+    authorize @saletypes
+  end
+
+  def new
+    # Create new instance of sale type
+    @saletype = SaleType.new
+    # Check with pundit if the user has permission
+    authorize @saletype
   end
 
   def create
+    # Create new instance of sale type from params
     @saletype = SaleType.new(sale_type_params)
+    # Check with pundit if the user has permission
+    authorize @saletype
+    # Survived so save
     if @saletype.save
       flash[:notice] = 'Sale type successfully created'
       redirect_to admin_sale_types_path
@@ -19,16 +32,19 @@ class Admin::SaleTypesController < ApplicationController
     end
   end
 
-  def new
-    @saletype = SaleType.new
-  end
-
   def edit
+    # Get sale type
     @saletype = SaleType.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @saletype
   end
 
   def update
+    # Get sale type
     @saletype = SaleType.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @saletype
+    # Survived so update
     if @saletype && @saletype.update(sale_type_params)
       flash[:notice] = 'Sale type successfully updated'
       redirect_to admin_sale_types_path
@@ -39,6 +55,9 @@ class Admin::SaleTypesController < ApplicationController
 
   def destroy
     @saletype = SaleType.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @saletype
+    # Survived so destroy
     if @saletype && @saletype.destroy
       flash[:notice] = 'Sale type successfully removed'
     else
