+++ app/controllers/admin/tenures_controller.rb
@@ -1,16 +1,29 @@
 class Admin::TenuresController < ApplicationController
   before_action :authenticate_user!
-  before_action :require_admin, except: [:index, :show]
 
   def index
+    # Get tenures
     @tenures = Tenure.all.paginate(
       :page => params[:page],
       :per_page => 10
     )
+    # Check with pundit if the user has permission
+    authorize @tenures
+  end
+
+  def new
+    # Create new instance
+    @tenure = Tenure.new
+    # Check with pundit if the user has permission
+    authorize @tenure
   end
 
   def create
+    # Create new instance from params
     @tenure = Tenure.new(tenure_params)
+    # Check with pundit if the user has permission
+    authorize @tenure
+    # Survived so save
     if @tenure.save
       flash[:notice] = 'Tenure successfully created'
       redirect_to admin_tenures_path
@@ -19,16 +32,19 @@ class Admin::TenuresController < ApplicationController
     end
   end
 
-  def new
-    @tenure = Tenure.new
-  end
-
   def edit
+    # Get tenure
     @tenure = Tenure.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @tenure
   end
 
   def update
+    # Get tenure
     @tenure = Tenure.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @tenure
+    # Survived so update
     if @tenure && @tenure.update(tenure_params)
       flash[:notice] = 'Tenure successfully updated'
       redirect_to admin_tenures_path
@@ -38,7 +54,11 @@ class Admin::TenuresController < ApplicationController
   end
 
   def destroy
+    # Get tenure
     @tenure = Tenure.find(params[:id])
+    # Check with pundit if the user has permission
+    authorize @tenure
+    # Survived so destroy
     if @tenure && @tenure.destroy
       flash[:notice] = 'Tenure successfully removed'
     else
